Improved the GroupManager configuration and renamed the package.
