+++ src/main/scala/mesosphere/marathon/api/v2/DeploymentsResource.scala
@@ -8,7 +8,7 @@ import javax.ws.rs.core.{ Context, MediaType, Response }
 
 import mesosphere.marathon.api.v2.json.Formats._
 import mesosphere.marathon.api.{ AuthResource, MarathonMediaType }
-import mesosphere.marathon.core.groupmanager.GroupManager
+import mesosphere.marathon.core.group.GroupManager
 import mesosphere.marathon.plugin.auth._
 import mesosphere.marathon.{ MarathonConf, MarathonSchedulerService }
 import mesosphere.util.Logging
