+++ src/main/scala/mesosphere/marathon/core/group/GroupManagerConfig.scala
@@ -1,13 +1,15 @@
-package mesosphere.marathon.core.groupmanager
+package mesosphere.marathon.core.group
 
 import org.rogach.scallop.ScallopConf
+
+import scala.concurrent.duration._
 
 trait GroupManagerConfig extends ScallopConf {
   //scalastyle:off magic.number
 
   lazy val groupManagerRequestTimeout = opt[Int](
     "group_manager_request_timeout",
-    descr = "Foo",
+    descr = "INTERNAL TUNING PARAMETER: Timeout (in ms) for requests to the group manager actor.",
     hidden = true,
-    default = Some(1000))
+    default = Some(10.seconds.toMillis.toInt))
 }
