--- src/main/scala/mesosphere/marathon/core/groupmanager/package.scala
@@ -1,6 +1,0 @@
-package mesosphere.marathon.core
-
-/**
-  * This package allows to manage all group related actions.
-  */
-package object groupmanager
