+++ src/test/scala/mesosphere/marathon/api/v2/AppVersionsResourceTest.scala
@@ -1,7 +1,7 @@
 package mesosphere.marathon.api.v2
 
 import mesosphere.marathon.api.TestAuthFixture
-import mesosphere.marathon.core.groupmanager.GroupManager
+import mesosphere.marathon.core.group.GroupManager
 import mesosphere.marathon.state.{ AppDefinition, Timestamp }
 import mesosphere.marathon.state.PathId._
 import mesosphere.marathon.test.Mockito
