+++ src/test/scala/mesosphere/marathon/api/v2/DeploymentsResourceTest.scala
@@ -1,7 +1,7 @@
 package mesosphere.marathon.api.v2
 
 import mesosphere.marathon.api.TestAuthFixture
-import mesosphere.marathon.core.groupmanager.GroupManager
+import mesosphere.marathon.core.group.GroupManager
 import mesosphere.marathon.state.{ AppDefinition, Group, PathId }
 import mesosphere.marathon.test.Mockito
 import mesosphere.marathon.upgrade.DeploymentManager.DeploymentStepInfo
