+++ src/test/scala/mesosphere/marathon/core/appinfo/impl/DefaultInfoServiceTest.scala
@@ -2,7 +2,7 @@ package mesosphere.marathon.core.appinfo.impl
 
 import mesosphere.marathon.MarathonSpec
 import mesosphere.marathon.core.appinfo.{ AppInfo, AppSelector, GroupInfo, GroupSelector }
-import mesosphere.marathon.core.groupmanager.GroupManager
+import mesosphere.marathon.core.group.GroupManager
 import mesosphere.marathon.state._
 import mesosphere.marathon.test.Mockito
 import org.scalatest.{ GivenWhenThen, Matchers }
